+++ onyx/crates/ember-http/src/status.rs
@@ -9,6 +9,7 @@ pub enum StatusCode {
     Unauthorized,
     Forbidden,
     NotFound,
+    MethodNotAllowed,
 
     InternalServerError,
 }
@@ -24,6 +25,7 @@ impl StatusCode {
             StatusCode::Unauthorized => 401,
             StatusCode::Forbidden => 403,
             StatusCode::NotFound => 404,
+            StatusCode::MethodNotAllowed => 405,
 
             StatusCode::InternalServerError => 500,
         }
@@ -39,6 +41,7 @@ impl StatusCode {
             StatusCode::Unauthorized => "Unauthorized",
             StatusCode::Forbidden => "Forbidden",
             StatusCode::NotFound => "Not Found",
+            StatusCode::MethodNotAllowed => "Method Not Allowed",
 
             StatusCode::InternalServerError => "Internal Server Error",
         }